:sparkles: exception 수정
