+++ src/main/java/com/team_3/nursing_care/common/exception/MemberException.java
@@ -1,0 +1,9 @@
+package com.team_3.nursing_care.common.exception;
+
+public class MemberException extends CustomException {
+
+    public MemberException(int code, String message) {
+        super(code, message);
+    }
+
+}
